making svg responsive and cleanup
finally, making the svg responsive so that it at least takes up a
correct amount of the screen. this involved the following steps:
* put the SVG in a container div whose width is some % (including media)
* remove the SVG width and height, and make a viewBox attribute instead
  setting the coordinate system
* remove the background image width and height, make them 100% instead
i think that was it actually

then, some additional modifications to the responsive columns idea. the
roll row requires some additional breakpoints, so while it should still
obey the 100% rule below 500px, the line is too long between 550px and
850px, and a pleasing amount of wrap is a little different for each
line. implemented and now it all flows super smoothly.